Cleaning up, adding comments/questions, adding information
Fixed coding omissions that made this not reproduce on other computers.

Added a couple sentences to intro section because I found the whole "bdi.8m" = 6 months post-followup to be rather confusing. Language added that clarifies the variable names.

Reformatted plots in variable description section so that they show up next to each other and more compact. Added multiplot function, which I found online, which allows this (facets do something different, as far as I can tell).

Changed plot titles to be consistent with one another. Fixed axis so that comparable plots have the same values.

Wrote up analysis of variables we discussed in class more or less similarly to Emily's writeup of the 3m variable. Added note that this part confuses me.

Added 2 hypotheses.

Added my comments to discussion about missing data. Put it in a chunk so it's more obvious it's commentary and not actually part of the writeup. Included the R code I used while contemplating the issue.

Put back in the original plot.

	modified:   BtheB4.Rnw
	modified:   BtheB4.pdf